refactor(core): isolate TLS and transport dependencies
