+++ crates/cli/Cargo.toml
@@ -23,12 +23,10 @@ clap.workspace = true
 
 # Everruns SDK
 everruns-sdk.workspace = true
+everruns-provider = { workspace = true, default-features = false, features = ["tls-ring"] }
 
 # HTTP client (for endpoints not yet in SDK)
 reqwest.workspace = true
-
-# TLS crypto provider
-rustls.workspace = true
 
 # Async runtime
 tokio.workspace = true
