+++ crates/cli/src/main.rs
@@ -169,9 +169,7 @@ pub enum CapabilitiesCommand {
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
-    // Install rustls crypto provider before any TLS usage.
-    // See everruns_core::telemetry::install_crypto_provider for the canonical helper.
-    let _ = rustls::crypto::ring::default_provider().install_default();
+    everruns_provider::install_ring_crypto_provider();
 
     let cli = Cli::parse();
     let output_format = output::OutputFormat::from_str(&cli.output);
