+++ crates/core/src/telemetry.rs
@@ -159,22 +159,6 @@ pub mod gen_ai {
 }
 
 // ============================================================================
-// Crypto Provider
-// ============================================================================
-
-/// Install the rustls crypto provider (ring) for TLS.
-///
-/// Must be called before any TLS usage. Without this, concurrent TLS
-/// connections (e.g. parallel tool execution in ActAtom) panic because
-/// rustls 0.23 cannot auto-detect the provider when multiple threads race.
-///
-/// Safe to call multiple times — subsequent calls are no-ops.
-pub fn install_crypto_provider() {
-    // Already-installed is expected if called from multiple init paths
-    let _ = rustls::crypto::ring::default_provider().install_default();
-}
-
-// ============================================================================
 // Span Helpers
 // ============================================================================
